#49 [mod] : 변수 타입 Long으로 변경

+++ app/src/main/java/com/byeboo/app/presentation/quest/QuestTipState.kt
@@ -6,8 +6,8 @@ import kotlinx.collections.immutable.persistentListOf
 
 @Immutable
 data class QuestTipState(
-    val stepNumber: Int = 0,
-    val questNumber: Int = 0,
+    val stepNumber: Long = 0,
+    val questNumber: Long = 0,
     val question: String = "연애에서 반복됐던 문제 패턴 3가지를 생각해보아요.",
     val tipStep: ImmutableList<Int> = persistentListOf(1, 2, 3),
     val tipQuestion: ImmutableList<String> = persistentListOf(
